+++ toolchain-linux.cmake
@@ -2,12 +2,18 @@
 #set(CMAKE_GENERATOR "Ninja")
 # Ninja doesn't support platform
 #set(CMAKE_GENERATOR_PLATFORM x64)
+if (NOT python)
+  set(python "python3") # default for linux
+endif()
 
-if (DEFINED ENV{ROCM_PATH})
-  set(rocm_bin "$ENV{ROCM_PATH}/bin")
-else()
-  set(ROCM_PATH "/opt/rocm" CACHE PATH "Path to the ROCm installation.")
-  set(rocm_bin "/opt/rocm/bin")
+
+if (NOT DEFINED ENV{ROCM_PATH})
+  set(ENV{ROCM_PATH} "/opt/rocm" CACHE PATH "Path to the ROCm installation.")
+endif()
+
+set(rocm_bin "$ENV{ROCM_PATH}/bin")
+if (NOT DEFINED CMAKE_PREFIX_PATH) 
+  list( APPEND CMAKE_PREFIX_PATH $ENV{ROCM_PATH}/llvm $ENV{ROCM_PATH})
 endif()
 
 if (NOT DEFINED ENV{CXX})